+++ src/Automation/Automation/ChangeLog.md
@@ -18,11 +18,12 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Fixed issue that disables the schedule for update deployment if schedule was re-created with same name
+* Added support for Customer Managed Key Encryption with System Assigned Identity
 
 ## Version 1.5.2
 * Fixed the issue for starting Python3 runbooks with parameters
 * Fixed time zone issue while creating an update deployment schedule.
+* Fixed issue that disables the schedule for update deployment if schedule was re-created with same name
 
 ## Version 1.5.1
 * Fixed te issue that complex object cannot be serialized correctly. [#14431]